+++ controllers/medicoController.js
@@ -1,6 +1,8 @@
 const Medico = require('../models/medico');
+const Auditoria = require('./../models/auditoria');
 const { Op } = require('sequelize');
-const path = require('path')
+const { sequelize } = require('../config/database');
+const jwt = require('jsonwebtoken');
 
 exports.getMedicos = async (req, res) => {
     try {
@@ -56,18 +58,54 @@ exports.getOneMedico = async (req, res) => {
 }
 
 exports.createMedico = async (req, res) => {
+    const t = await sequelize.transaction();
+
+    const token = req.headers.authorization?.split(' ')[1];
+    if (!token) {
+        return res.status(401).json({
+            status: 'fail',
+            message: 'Token de autenticação não fornecido.'
+        });
+    }
+
+    const decode = jwt.decode(token);
+    if (!decode || !decode.usuario) {
+        return res.status(404).json({
+            status: 'fail',
+            message: 'Usuário não encontrado.'
+        });
+    }
+
+    const data = new Date();
+     
     try { 
+        await Auditoria.create({
+            entidade: 'medicos',
+            acao: 'create',
+            dados_novos: req.body,
+            usuario_id: decode.id,
+            data: data
+        }, {
+            transaction: t
+        })
+
         const medico = await Medico.create({
             nome: req.body.nome,
             crm: req.body.crm,
             telefone: req.body.telefone,
             email: req.body.email
+        }, {
+            transaction: t    
         })
+
+        await t.commit();
+
         res.status(201).json({
             status: "success",
             data: medico
 
     }) } catch (err) { 
+        await t.rollback();
         res.status(400).json ({
             status: "fail",
             message: err.message
@@ -76,25 +114,64 @@ exports.createMedico = async (req, res) => {
 }
 
 exports.updateMedico = async (req, res) => {
+
+    const t = await sequelize.transaction();
+
+    const token = req.headers.authorization?.split(' ')[1];
+    if (!token) {
+        return res.status(401).json({
+            status: 'fail',
+            message: 'Token de autenticação não fornecido.'
+        });
+    }
+
+    const decode = jwt.decode(token);
+    if (!decode || !decode.usuario) {
+        return res.status(404).json({
+            status: 'fail',
+            message: 'Usuário não encontrado.'
+        });
+    }
+
+    const data = new Date();
+
     try {
-        const resultado = await Medico.update(req.body, {
-            where: {
-                crm: req.params.crm
-            }
+
+        const registroAnterior = await Medico.findByPk(req.params.crm, {
+            transaction: t
+        });
+        
+        await Auditoria.create({
+            entidade: 'medicos',
+            acao: 'update',
+            dados_anteriores: registroAnterior,
+            dados_novos: req.body,
+            usuario_id: decode.id,
+            data: data
+        }, {
+            transaction: t
         })
 
-        if (resultado === 0) {
+        const [updatedRows] = await Medico.update(req.body, {
+            where: { crm: req.params.crm },
+            transaction: t
+        });
+        
+        if (updatedRows === 0) {
             return res.status(404).json({
                 status: 'fail',
-                message: 'Nenhum paciente encontrado com o CPF fornecido.'
+                message: 'Nenhum médico encontrado com o CRM fornecido.'
             });
         }
+
+        await t.commit();
 
         res.status(200).json({
             status: 'success',
             message: 'Médico alterado com sucesso.'
         })
     } catch (err) {
+        await t.rollback();
         res.status(400).json({
             status: 'fail',
             message: err.message
@@ -103,25 +180,65 @@ exports.updateMedico = async (req, res) => {
 }
 
 exports.deleteMedico = async (req, res) => {
+    const t = await sequelize.transaction();
+
+    const token = req.headers.authorization?.split(' ')[1];
+    if (!token) {
+        return res.status(401).json({
+            status: 'fail',
+            message: 'Token de autenticação não fornecido.'
+        });
+    }
+
+    const decode = jwt.decode(token);
+    if (!decode || !decode.usuario) {
+        return res.status(404).json({
+            status: 'fail',
+            message: 'Usuário não encontrado.'
+        });
+    }
+
+    const data = new Date();
+
     try {
+
+        const registroAnterior = await Medico.findByPk(req.params.crm, {
+            transaction: t
+        })
+
+        await Auditoria.create({
+            entidade: 'medicos',
+            acao: 'create',
+            dados_anteriores: registroAnterior,
+            usuario_id: decode.id,
+            data: data
+        }, {
+            transaction: t
+        })
+
         const resultado = await Medico.destroy({
             where: {
                 crm: req.params.crm
-            }
+            }, 
+            transaction: t
         })
 
         if (resultado === 0) {
-            res.status(404).json({
+            return res.status(404).json({
                 status: 'fail',
                 message: 'Nenhum médico foi encontrado'
             })
         }
 
+        await t.commit();
+
         res.status(200).json({
             status: 'success',
             message: 'Médico deletado com sucesso'
         })
+
     } catch (err) {
+        t.rollback();
         res.status(404).json({
             status: 'fail',
             message: err.message
